% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewens_k_exact.R
\name{ewens_k_exact}
\alias{ewens_k_exact}
\title{Calculate expected number of classes in a sample of size n given theta}
\usage{
ewens_k_exact(n, theta)
}
\arguments{
\item{n}{The sample size}

\item{theta}{The non-negative parameter governing expected sample diversity}
}
\description{
The expected number of classes from the Ewens distribution is given by \eqn{\theta \sum_{j=1}^{n} \frac{1}{\theta + j - 1}}. This is often more convenient than integrating across the PMF given by \code{dewens_k}
}
