% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewens_mle.R
\name{ewens_mle}
\alias{ewens_mle}
\title{Maximum likelihood estimate of theta given sample vector with class memberships}
\usage{
ewens_mle(x)
}
\arguments{
\item{x}{A vector containing class memberships; sample size n and number of classes k are calculated from this}
}
\description{
Maximum likelihood estimate of theta given sample vector with class memberships
}
