% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcrp.R
\name{gcrp}
\alias{gcrp}
\title{Draw from a generalized Chinese Restaurant Process}
\usage{
gcrp(n, alpha = 0, theta = 1)
}
\arguments{
\item{n}{The sample size.}

\item{alpha}{A parameter between zero and one inclusive governing the expected sample diversity}

\item{theta}{A non-negative parameter governing the expected sample diversity.}
}
\value{
A vector of length n consisting of numeric class labels.
}
\description{
Draw from a generalized Chinese Restaurant Process
}
\examples{
rewens(100, 1)
rewens(120, 0.5)
rewens(10, 0)
}
