% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewens.R
\name{dewens}
\alias{dewens}
\title{Probability mass function for the Ewens distribution}
\usage{
dewens(x, theta = 1, log = FALSE)
}
\arguments{
\item{x}{A vector giving class memberships of each observation in the sample}

\item{theta}{A non-negative parameter governing the expected sample diversity.}

\item{log}{if TRUE, probabilities are given as log(p)}
}
\description{
Probability mass function for the Ewens distribution
}
\examples{
x <- sample(LETTERS, 120, replace = TRUE)
dewens(x, theta = 1)
dewens(x, theta = 0) ## returns NaN since vector incompatible with zero diversity

}
