% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewens_k.R
\name{dewens_k}
\alias{dewens_k}
\title{Probability mass function for the number of classes from a Ewens distribution}
\usage{
dewens_k(k, n, theta)
}
\arguments{
\item{k}{An integer number of classes at which to evaluate the PMF}

\item{n}{A sample size not less than k}

\item{theta}{A non-negative parameter governing the expected sample diversity.}
}
\description{
Probability mass function for the number of classes from a Ewens distribution
}
\examples{
x <- sample(LETTERS, 120, replace = TRUE)
dewens_k(1, 20, theta = 1)

}
