% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewens_k_exact.R
\name{ewens_k_exact}
\alias{ewens_k_exact}
\title{Calculate expected number of classes in a sample of size n given theta}
\usage{
ewens_k_exact(n, theta)
}
\arguments{
\item{n}{The sample size}

\item{theta}{The non-negative parameter governing expected sample diversity}
}
\description{
Calculate expected number of classes in a sample of size n given theta
}
