% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewens.R
\name{rewens}
\alias{rewens}
\title{Draw from the Ewens distribution}
\usage{
rewens(n, theta = 1)
}
\arguments{
\item{n}{The sample size.}

\item{theta}{A non-negative parameter governing the expected sample diversity.}
}
\value{
A vector of length n consisting of numeric class labels.
}
\description{
Draw from the Ewens distribution
}
\examples{
rewens(100, 1)
rewens(120, 0.5)
rewens(10, 0)
}
