% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgem.R
\name{rgem}
\alias{rgem}
\title{Draw from the Griffiths-Engen-McCloskey distribution}
\usage{
rgem(alpha = 0, theta = 1, trunc_at = 500)
}
\arguments{
\item{alpha}{A parameter between zero and one}

\item{theta}{A parameter which must be greater than -alpha}

\item{trunc_at}{An integer which specifies the maximum number of components to return}
}
\value{
A vector of shares of length \code{trunc_at} which may sum to less than one
}
\description{
Draw from the Griffiths-Engen-McCloskey distribution
}
